+++ examples/alert_source/variables.tf
@@ -12,8 +12,3 @@ variable "password" {
   description = "iLert password used to configure the provider"
   type        = string
 }
-
-variable "escalation_policy_id" {
-  description = "iLert escalation policy id used for alert source"
-  type        = number
-}