+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## 20.10.2020, Version 1.0.0
 
 - add alert source data source and resource
+- add escalation policy data source and resource
+- add schedule data source and resource
+- add user data source and resource